Update lecture file inside, outside
