+++ packages/inside/src/styles/lecture/detail/style.ts
