+++ shared/common/src/pages/index.ts
@@ -1,6 +1,5 @@
 export { default as LoginPage } from './login'
 export { default as SignUpPage } from './signUp'
 export { default as HomePage } from './home'
-export { default as DetailPage } from './lecture/detail'
 export { default as ClubPage } from './club'
 export { default as NoticePage } from './notice'
